Added start menu
