+++ src/index.ejs
@@ -8,6 +8,17 @@
 <body>
     <div id="gameArea">
         <canvas id="gameCanvas"></canvas>
+        <div id="gameMenu">
+            <div id="title">
+                Tanky McTankface
+            </div>
+            <div id="buttonArea">
+                    <button class="button" id="startbutton">Start</button>
+            </div>
+            <div id="footer">
+                A <a href="https://js13kgames.com/">js13k Game</a> - © 2019 - Frederik Petersen
+            </div>
+        </div>
     </div>
 </body>
 </html>
