+++ src/main.css
@@ -10,9 +10,57 @@
     height: 100%;
 }
 
+#gameMenu {
+    position: absolute;
+    z-index: 1;
+    width: 100%;
+    height: 100%;
+    bottom: 0;
+    opacity: 0.8;
+    background-color: #000000;
+}
+
+#buttonArea {
+    padding-top: 10%;
+    text-align: center; 
+}
+
+.button {
+    background-color: yellow;
+    border: none;
+    color: black;
+    padding: 15px 32px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 2em;
+  }
+
+.button:hover {
+    background-color: greenyellow;
+}
+
+#title {
+    text-align: center;
+    padding-top: 10%;
+    font-size: 4em;
+    color: yellow;
+}
+
+#footer {
+    position: absolute;
+    text-align: center;
+    color: white;
+    bottom: 0;
+    font-size: 1.5em;
+    width: 100%;
+    height: 20%;
+}
+
 html, body {
     height: 100%;
     overflow: hidden;
     margin: 0;
     background-color: #000000;
-  }
+    font-family: monospace;
+}
